Fix DLGTEMPLATE and DLGITEMTEMPLATE field sizes

+++ src/User32/User32+DLGITEMTEMPLATE.cs
@@ -21,40 +21,40 @@ namespace PInvoke
             ///     The style of the control. This member can be a combination of window style values (such as WS_BORDER) and one or
             ///     more of the control style values (such as BS_PUSHBUTTON and ES_LEFT).
             /// </summary>
-            public int style;
+            public uint style;
 
             /// <summary>
             ///     The extended styles for a window. This member is not used to create controls in dialog boxes, but applications that
             ///     use dialog box templates can use it to create other types of windows.
             /// </summary>
-            public int dwExtendedStyle;
+            public uint dwExtendedStyle;
 
             /// <summary>
             ///     The x-coordinate, in dialog box units, of the upper-left corner of the control. This coordinate is always relative
             ///     to the upper-left corner of the dialog box's client area.
             /// </summary>
-            public byte x;
+            public short x;
 
             /// <summary>
             ///     The y-coordinate, in dialog box units, of the upper-left corner of the control. This coordinate is always relative
             ///     to the upper-left corner of the dialog box's client area.
             /// </summary>
-            public byte y;
+            public short y;
 
             /// <summary>
             ///     The width, in dialog box units, of the control.
             /// </summary>
-            public byte cx;
+            public short cx;
 
             /// <summary>
             ///     The height, in dialog box units, of the control.
             /// </summary>
-            public byte cy;
+            public short cy;
 
             /// <summary>
             ///     The control identifier.
             /// </summary>
-            public short id;
+            public ushort id;
         }
     }
 }
